Lint Filters, Fixed Api Path, Deleted Connection file, updated Schema & Seeds.sql
