+++ config/config.json
@@ -1,14 +1,14 @@
 {
   "development": {
     "username": "root",
-    "password": null,
-    "database": "passport_demo",
+    "password": "password",
+    "database": "easy_serve",
     "host": "127.0.0.1",
     "dialect": "mysql"
   },
   "test": {
     "username": "root",
-    "password": null,
+    "password": "password",
     "database": "database_test",
     "host": "127.0.0.1",
     "dialect": "mysql"
