+++ routes/index.js
@@ -1,10 +1,8 @@
-const router = require("express").Router();
-const apiRoutes = require("./api");
-const htmlRoutes = require("./html");
-const connection = require("./api/connection");
+const router = require('express').Router();
+const apiRoutes = require('./api');
+const htmlRoutes = require('./html');
 
-router.use("/data", connection);
-router.use("/api", apiRoutes);
-router.use("/", htmlRoutes);
+router.use('/api', apiRoutes);
+router.use('/', htmlRoutes);
 
 module.exports = router;
